+++ tracker/statstracker_test.go
@@ -1,4 +1,4 @@
-package logger
+package tracker
 
 import (
 	"testing"
